added extension point to allow people to not use the shitty MS javascript serializer
