+++ src/AspComet/ISerializer.cs
@@ -1,0 +1,8 @@
+﻿namespace AspComet
+{
+    public interface ISerializer
+    {
+        string Serialize(object obj);
+        T Deserialize<T>(string json);
+    }
+}
